\name{annotation}
\alias{annotation}

\title{Annotations of the RAS pathway related to dataset}
\usage{data(annotation)}
\arguments{
  \item{annotation}{Consist of two columns, first refers gene symbols and second refers gene symbol. }
}
\value{
  \code{annotation} returns annotation data.
}
\description{
  \code{annotation} Subset annotation data of Affymetrix HG-U95AV2 GeneChip.
}
\details{
      The microarray platform used in the prostate cancer study is the Affymetrix HG-U95AV2 GeneChip.
 }


\examples{
data(annotation)
}
\keyword{misc}

