\name{dataset}
\alias{dataset}

\title{Example data set}
\usage{data(dataset)}
\arguments{
  \item{dataset}{Example data rows correspond to variables (e.g. genes) and columns to samples.}
}
\value{
  \code{dataset} returns mRNA expression dataset
}
\description{
  \code{dataset} This dataset contains the gene expression, annotations and clinical data as published by Jorissen and colleagues in 2009. 
}

\references{
      
	http://www.ncbi.nlm.nih.gov/projects/geo/query/acc.cgi?acc=GSE14333
}


\examples{
data(dataset)
}
\keyword{misc}

