\name{omim}
\alias{omim}

\title{OMIM diseases gene list which contains breast cancer related data}
\usage{data(omim)}
\arguments{
  \item{omim}{Gene list}
}
\value{
  \code{omim} returns omim gene list.
}
\description{
  \code{omim} OMIM diseases gene list which contains breast cancer related data. It was obtained via Enrichr tool
}
\details{
      OMIM diseases gene list which contains breast cancer related data
 }


\examples{
data(omim)
}
\keyword{misc}

