\name{ppi}
\alias{ppi}

\title{Example literature dataset}
\usage{data(ppi)}
\arguments{
  \item{ppi}{Protein protein interactions}
}
\value{
  \code{ppi} returns protein protein interactions
}
\description{
  \code{ppi} This ppi is utilized as literature data in overlap analysis. It contains more than 1.5 million protein protein interactions which can be access by GAnet R package.
}

\examples{
data(ppi)
}
\keyword{misc}

